Refactor: 패키지 구조 계층형으로 변경 [#4]
