+++ src/main/java/com/hrbank/hrbank/dto/CursorPageResponseDto.java
@@ -1,4 +1,4 @@
-package com.hrbank.hrbank.domain.department.dto;
+package com.hrbank.hrbank.dto;
 
 import java.util.List;
