+++ src/main/java/com/hrbank/hrbank/dto/DepartmentDto.java
@@ -1,4 +1,4 @@
-package com.hrbank.hrbank.domain.department.dto;
+package com.hrbank.hrbank.dto;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
