+++ src/main/java/com/hrbank/hrbank/repository/DepartmentRepository.java
@@ -1,6 +1,6 @@
-package com.hrbank.hrbank.domain.department.repository;
+package com.hrbank.hrbank.repository;
 
-import com.hrbank.hrbank.domain.department.entity.Department;
+import com.hrbank.hrbank.entity.Department;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
